Add fly and higher jump powerup; Made changes in PlayerMovement script
Player was able to "double dodge" to the edge lanes, and this resets the Lerp process, what made the movement slow. Now, the player is able to "cancel the dodge" (Ex: He moves to the left edge lane, and while moving, he can back to the previous lane), but he's no able to double dodge anymore.