Update PlayerController.cs
Now, the script "stores" the next movement in the memory. Ex: If the player is dodging and swipe up the finger on the screen (aka press de jump button), the game will wait the player finish the dodge movement and, after that, will automatically apply the jump movement. The last input pressed is stored in the memory.